api/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-API==3.1
Flask-MySQLdb==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
